waycode/requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
numpy>=1.24.0
click>=8.1.0